expression: make some unstable test serial
